修改bug剔除很多 文件

+++ app/src/main/java/com/honliv/honlivmall/bean/DefaultPrice.java
@@ -3,18 +3,18 @@ package com.honliv.honlivmall.bean;
 import java.io.Serializable;
 
 public class DefaultPrice implements Serializable {
-	private int minPrice;
-	private int maxPrice;
-	public int getMinPrice() {
+	private float minPrice;
+	private float maxPrice;
+	public float getMinPrice() {
 		return minPrice;
 	}
-	public void setMinPrice(int minPrice) {
+	public void setMinPrice(float minPrice) {
 		this.minPrice = minPrice;
 	}
-	public int getMaxPrice() {
+	public float getMaxPrice() {
 		return maxPrice;
 	}
-	public void setMaxPrice(int maxPrice) {
+	public void setMaxPrice(float maxPrice) {
 		this.maxPrice = maxPrice;
 	}
 	@Override
